requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
unidecode
google-cloud-bigquery
db-dtypes
pyarrow
